chore(quarkus): Factorize Camel catalog runtime generation logic
