+++ pkg/trait/camel.go
@@ -19,13 +19,7 @@ package trait
 
 import (
 	"fmt"
-	"io/ioutil"
-	"os"
-	"path"
-	"regexp"
 	"strings"
-
-	yaml2 "gopkg.in/yaml.v2"
 
 	"github.com/pkg/errors"
 
@@ -33,9 +27,6 @@ import (
 
 	"github.com/apache/camel-k/pkg/apis/camel/v1alpha1"
 	"github.com/apache/camel-k/pkg/util/camel"
-	"github.com/apache/camel-k/pkg/util/defaults"
-	"github.com/apache/camel-k/pkg/util/kubernetes"
-	"github.com/apache/camel-k/pkg/util/maven"
 )
 
 type camelTrait struct {
@@ -112,17 +103,8 @@ func (t *camelTrait) loadOrCreateCatalog(e *Environment, camelVersion string, ru
 		// if the catalog is not found in the cluster, try to create it if
 		// the required versions (camel and runtime) are not expressed as
 		// semver constraints
-		cvHasFixedVersion, err := regexp.MatchString(`^(\d+)\.(\d+)\.([\w-\.]+)$`, camelVersion)
-		if err != nil {
-			return err
-		}
-		rvHasFixedVersion, err := regexp.MatchString(`^(\d+)\.(\d+)\.([\w-\.]+)$`, runtimeVersion)
-		if err != nil {
-			return err
-		}
-
-		if cvHasFixedVersion && rvHasFixedVersion {
-			c, err = t.generateCatalog(e, camelVersion, runtimeVersion)
+		if exactVersionRegexp.MatchString(camelVersion) && exactVersionRegexp.MatchString(runtimeVersion) {
+			c, err = camel.GenerateCatalog(e.C, e.Client, ns, e.Platform.Spec.Build.Maven, camelVersion, runtimeVersion)
 			if err != nil {
 				return err
 			}
@@ -148,90 +130,6 @@ func (t *camelTrait) loadOrCreateCatalog(e *Environment, camelVersion string, ru
 	e.CamelCatalog = c
 
 	return nil
-}
-
-func (t *camelTrait) generateCatalog(e *Environment, camelVersion string, runtimeVersion string) (*camel.RuntimeCatalog, error) {
-	root := os.TempDir()
-	tmpDir, err := ioutil.TempDir(root, "camel-catalog")
-	if err != nil {
-		return nil, err
-	}
-
-	defer os.RemoveAll(tmpDir)
-
-	if err := os.MkdirAll(tmpDir, os.ModePerm); err != nil {
-		return nil, err
-	}
-
-	project := t.generateMavenProject(camelVersion, runtimeVersion)
-
-	mc := maven.NewContext(tmpDir, project)
-	mc.LocalRepository = e.Platform.Spec.Build.Maven.LocalRepository
-	mc.Timeout = e.Platform.Spec.Build.Maven.Timeout.Duration
-	mc.AddSystemProperty("catalog.path", tmpDir)
-	mc.AddSystemProperty("catalog.file", "catalog.yaml")
-
-	ns := e.DetermineNamespace()
-	if ns == "" {
-		return nil, errors.New("unable to determine namespace")
-	}
-
-	settings, err := kubernetes.ResolveValueSource(e.C, e.Client, ns, &e.Platform.Spec.Build.Maven.Settings)
-	if err != nil {
-		return nil, err
-	}
-	if settings != "" {
-		mc.SettingsContent = []byte(settings)
-	}
-
-	err = maven.Run(mc)
-	if err != nil {
-		return nil, err
-	}
-
-	content, err := ioutil.ReadFile(path.Join(tmpDir, "catalog.yaml"))
-	if err != nil {
-		return nil, err
-	}
-
-	catalog := v1alpha1.CamelCatalog{}
-	if err := yaml2.Unmarshal(content, &catalog); err != nil {
-		return nil, err
-	}
-
-	return camel.NewRuntimeCatalog(catalog.Spec), nil
-}
-
-func (t *camelTrait) generateMavenProject(camelVersion string, runtimeVersion string) maven.Project {
-	p := maven.NewProjectWithGAV("org.apache.camel.k.integration", "camel-k-catalog-generator", defaults.Version)
-
-	p.Build = &maven.Build{
-		DefaultGoal: "generate-resources",
-		Plugins: []maven.Plugin{
-			{
-				GroupID:    "org.apache.camel.k",
-				ArtifactID: "camel-k-maven-plugin",
-				Version:    runtimeVersion,
-				Executions: []maven.Execution{
-					{
-						ID: "generate-catalog",
-						Goals: []string{
-							"generate-catalog",
-						},
-					},
-				},
-				Dependencies: []maven.Dependency{
-					{
-						GroupID:    "org.apache.camel",
-						ArtifactID: "camel-catalog",
-						Version:    camelVersion,
-					},
-				},
-			},
-		},
-	}
-
-	return p
 }
 
 func (t *camelTrait) determineCamelVersion(e *Environment) string {
